+++ @types/globals.d.ts
@@ -58,3 +58,10 @@ interface Window {
   String: typeof String;
   wrappedJSObject?: Window;
 }
+
+// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
+interface Navigator {
+  keyboard?: {
+    getLayoutMap: () => Promise<Iterable<readonly [string, string]>>;
+  };
+}
